working version of 128 bit
There is a multiplication at PLA under cross function
If the kmer is at 126 bit (k=63), this multiplication can
result in >128 bit, which would result in an integer
overflow.

For both basic and repeat pla if this is the case,
this would not work.

For repeat-pla, the y-intersection can be very large
(as it is calculated on the next k-mer).
Thus, repeat-pla will have a lesser leeway for k-mer
value tolerance.

Otherwise, this works.